requirements.txt
# Core dependencies
openai>=1.0.0
streamlit>=1.28.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Document processing
pdfplumber>=0.11.0
python-docx>=0.8.11

# LangChain
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# Vector databases
faiss-cpu>=1.8.0

# HTTP and async
aiohttp>=3.8.0
requests>=2.25.0
httpx>=0.24.0

# Data processing
feedparser>=6.0.0
pandas>=1.3.0
numpy>=1.21.0
plotly>=5.0.0

# Database (optional)
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
